The project now uses dependency injection container.
-Moved the project config and database instances into dependency container instead of saving it globally on a variable.
-The container got 3 methods bind, unbind and resolve.

bind(dpName, dpValue) => binds the new dependency into a static associative array with dpName as the key and dpValue as the value of the key.

unbind(dpName) => unset the dpName from the static dependency variable.

resolve(dpName) => returns a value of the dpName if found, if not throws an exception.